requirements.txt
streamlit==1.37.0
pandas==2.2.2
plotly==5.22.0
matplotlib==3.10.3
numpy
google-api-python-client==2.182.0
google-auth==2.40.3